Fixed bug with quadtree. This means that quadtree lighting and rendering now work correctly. Unfortunatelly this also means that this is slow. VEEERRRYYY slow. I will try to fix this ASAP. New 'quadtree' console command to enable quadtree opt instead of c-buffer. This is still experimental.
Fixed bug in dynamic lightmap reading (pseudo-dynamic lights). This
bug only occurs in rare situations.

Fixed collision detection for infinite maze. This means no more crashes
for infinite maze but CD is still weird. We need a fly mode for the
infinite maze.

Changed lighting routines so that they no longer use the transformation
manager. Not only is this more efficient but it also fixes infinite
maze so that there are no more ugly glitches when new sectors
are created.


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@1825 8cc4aa7f-3514-0410-904f-f2cc9021211c