Reverted Paul's changes to the volatile.h generation code in the makefiles for several reasons:
  o He changed the semantic meaning of MAKE_VOLATILE_H, making it
    less flexible, order dependent, and less robust.  That is, after
    his change, one had to be very careful about the context in
    which MAKE_VOLATILE_H was used, whereas previously no such
    restriction applied.

  o He added a large section of copy/paste code to mingw.mak which
    unnecessarily increased maintenance costs and fragility.

  o Most of the changes were simply unnecessary.  It appears that
    much of the trouble was caused by a misplaced definition of $".

My fixes were twofold.  Fist I corrected the $" problem so that it is in
scope in the proper contexts.  Second, I further generalized and
abstracted the actual emission of volatile.h so that platform-specific
makefiles can intercede at all points of the process, if necessary.

Having accomplished this, I added very simple special-case code to
mingw.mak to perform the transformations on volatile.h which Paul
required.  This allowed me to eliminate all the copy/paste code which he
had added while still retaining flexibility.


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@4006 8cc4aa7f-3514-0410-904f-f2cc9021211c