Reverted Philip's changes to blocks.cpp.  Philip's changes did not solve the problem.  His changes merely side-stepped it by leaking the csSystemDriver.  This is a case of addressing the symptom rather than the cause.
Philip, try to narrow down the problem further.  Based upon your attempts
thus far, it appears that you have narrowed it down to the destruction of
the system driver.  However, system driver destruction is quite involved,
and a lot of code gets executed during that process.  If you could break
into the debugger and obtain a back-trace listing at the poing where the
application hangs, it could help in further narrowing down the problem.
Other options are to step through the ~Block() destructor as well as the
~csSystemDriver() destructor in order to locate the statement or function
call which is hanging.  Then step through the function which is hanging
and see which of its statements hangs.  Perform this process recursively
until you have found the offending code.

I also took this opportunity to transmogrify the `view' variable into an
instance variable of the Blocks class.  It used to be global.


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@6365 8cc4aa7f-3514-0410-904f-f2cc9021211c