Added a bumpmapping test application. It contains a procedural texture that is used to approximate bumpmapping. 3d hardware may be able to do bumpmapping nicer looking and faster, if it supports bumpmapping. If it does not support bumpmapping, the procedural texture is a good fallback option, that will run on almost any hardware (and software too if we fix that).
Note that the software renderer does not really like the tiling and
DPFX calls, so the texture is stretched a bit - there is one triangle
where you can see what it should look like. Under opengl the results
look nice.

Note that this is a test for now. The code has not been optimized at all.
And the computed overlays could be cached too. So don't look at
framerates.

When you start the application, you can see some wood texture on the
left (unbumpmapped). You can see a stone texture on the right (the
bumpmap texture). And the wood bumpmapped with stone is in the middle.
In the lower right hand corner you can see a debug view of the bumpmapping.


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@6485 8cc4aa7f-3514-0410-904f-f2cc9021211c