Due to unfortunate wording I have unintentionally caused harm to Philip's good name.  I would like to rectify
that situation by clarifying the meaning of the commit log posted earlier in which I referred to "damage."

The "damage" I was referring to was the accidental removal and/or modification of non-malignant code
structures due to a self-confessed failure to properly update the CVS tree.  The modifications "damaged" the
library in that, when properly synchronized, the "repaired" source code files no longer functioned properly
with code that utilized the expanded functionality of the affected classes.  Additional side effects related
to this also caused the code to be "damaged" in that it no longer appeared properly.

I was not attempting to demean or in anyway discredit Philip, as it is admittedly
his responsibility to make sure that certain things work.  In addition, the changes he made to the code were
beneficial because they allowed a non-compilable module to be compilable again.

I'm sorry if the post seemed rude or inconsiderate, or seemed to cast the blame in anyway upon Philip.  Pleaseunderstand that any blame for the CSWS compile failure rests entirely upon my shoulders. In the future I will
try to be clearer and use less inflammatory wording.

-={C}=-


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@6609 8cc4aa7f-3514-0410-904f-f2cc9021211c