Changed the directory structure of where VC outputs files.
For example for a CS build you will now see the new structure:


 CS\mk\visualc\csdebug (OR) csrelease
                   |
                   |-- bin
                   |    |
                   |    |-- libs
                   |
                   |-- temp


Now all ".lib" files are placed in "bin\libs"

Now all ".exe" and ".dll" files via a new "copy to", are
copied to the "bin" directory as well as the default "CS\"
root directory.

All object and VC internmediate files are created sub of "temp".

This change will allow, the much easier collection and copying
of files and/or working with both a debug or a release build of
CS via copy and pastes to the "CS\" root directory and or setting
of enviroment variables to use these directories as working ones.
Also makes it easier for file collection for own projects and
possible binary releases of the CS files.

*** WARNING ***

If you compile both debug and release builds of CS you must expect
the disk space used on your system to be around "1.1Gb" and the
the possibilty of mixing debug and release files inside the "CS\"
root directory is possible, so must be done only with care...

*** IMPORTANT ***

The following commit requires you delete the folder(s):

"CS\mk\visualc\Debug" and/or "CS\mk\visualc\Release"

and perform a full rebuild of the CS tree.


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@6734 8cc4aa7f-3514-0410-904f-f2cc9021211c