The awsTextureManager is now completely functional, if untested, other than the register/unregister functionality.  Once I figure out how to unregister a texture from the texture manager, that will work as well.
I still need to make some updates to the AwsPrefs class so that it will expose the functionality of the texture manager.  With this update I can now begin to actually test drawing code, since nearly all of the AWS enhancements require access to bitmaps for background textures and for bitmappable buttons/etc.


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@7684 8cc4aa7f-3514-0410-904f-f2cc9021211c