Updated winlibs version number to 0.19h.
Note this version of winlibs is the same as 0.19g but the
zlib and png libs and dll's have been built with VC6 using
the /LINK50COMPAT so they are again VC5 compatible.

The new version of winlibs can be obtained via the
Crystal Space documentation under MSVC specific build
instructions - where links to the file on both the CS ftp site
and my own mirror can be found.


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@8156 8cc4aa7f-3514-0410-904f-f2cc9021211c