Eric Sunshine added a new templated scfInterface<> class to scf.h.  This class can be used to perform queries for static information relating to SCF interfaces.  The following information is supplied by this class:
GetVersion()
    Retrieves an interface's current version number.  This is
    the same as the value of the global constant name_VERSION,
    where `name' is the interface's name (such as iVFS_VERSION).

GetID()
    Retrieves an interface's unique identifier as an
    scfInterfaceID.  This is the same as the value returned by
    the global function name_scfGetID(), where `name' is the
    interface's name (such as iVFS_scfGetID()).

As an example of use, to retrieve the version number of the
iFooBar interface, the following code can be used:

    int ver = scfInterface<iFooBar>::GetVersion();

An important benefit of scfInterface<> is that it allows
template authors to retrieve static information about an
interface even though the template might know an interface only
by the opaque tag `T'.  For instance
scfInterface<T>::GetVersion() will work as expected from within
a template class.


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@15782 8cc4aa7f-3514-0410-904f-f2cc9021211c