Added sigslot implementation from sigslot.sourceforge.net.  This is actually the fixed version from a link on the forums that actually compiles with recent compilers.
Also added two signals to the properties.  You can subscribe to both the Binding and Changed signals.  The Binding signal is fired whenever a property is bound to a given name.  (A property may only be bound to one name at a time.)

The Changed signal fires each time a successful Set operation occurs.  Trying to Set a read-only property never results in signal generation.

In this way it is simple for a programmer to know when changes to any given property occur.


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@21196 8cc4aa7f-3514-0410-904f-f2cc9021211c