AWS2: Changed the default storage in the autom::string object to use csString instead of std::string.  This is the last large std::string changeover, other than the parser.  There are a few instances of std::string that probably do no harm and are used here and there.  I may change these as I come to them.
csString: I added a constructor that allows you to specify a C-string (null terminated or not) and the length of that string.  This is safer than just constructing with a C-string because it requires the app to know how big the string is, and doesn't overread it's bounds like strlen might.


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@21777 8cc4aa7f-3514-0410-904f-f2cc9021211c