Eric Sunshine made the following changes:
Changed 'MasterHeader' rule in helper.jam so that it invokes the
'ResponseFile' rule with a delimiter of "%" rather than allowing
'ResponseFile' to use the default delimiter of "@", since "@" may legally
appear in email addresses and Doxygen directives (@file, for example),
both of which are likely to appear in master header boilerplate.

Changed CS/include/Jamfile so that it uses Doxygen directive "@file" in
master header boilerplate rather than "\file" since "\file" ends up as
"<formfeed>ile" when 'jam masterheaders' is invoked under the Dash
shell. (This appears to be a bug in Dash single-quote treatment and has
been reported.)


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@28245 8cc4aa7f-3514-0410-904f-f2cc9021211c