Eric Sunshine made the following changes:
Fixed a couple bugs in macro CS_EMIT_BUILD_PROPERTY() in emit.m4 which
manifested when the client passed a non-empty EMPTY-OKAY argument. (1) It
incorrectly unconditionally employed the CS_JAMCONFIG_PROPERTY() emitter
rather than respecting the EMITTER argument and only employing
CS_JAMCONFIG_PROPERTY() as a fallback. (2) It neglected to pass the
optional UNCONDITIONAL argument along to the emitter.

Added "default" as an alias for the existing tokens "emit" and "yes" for
the EMITTER argument of macro CS_EMITTER_OPTIONAL() in emit.m4.

Made minor clarifications to macro documentation in emit.m4.


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@28440 8cc4aa7f-3514-0410-904f-f2cc9021211c