Eric Sunshine made the following changes:
Added a check to configure.ac for GCC's __attribute__((init_priority(X))),
which is used by csutil/ptmalloc_wrap.cpp, and which is not universally
available (for instance, gcc 3.4.3 on OpenSolaris lacks it).

Added preprocessor macro CS_ATTRIBUTE_INIT_PRIORITY(), which is defined
appropriately for compilers which support the attribute, else is a no-op
for those which do not.

Slightly simplified __attribute__((malloc)) check in configure.ac.

Augmented CS_SYMBOL_QUALIFIER() in qualify.m4 so that it appends the
attribute to SYMBOL rather than prepending it.  In addition to being a
more canonical arrangement, this makes it possible to supply complex
SYMBOL definitions, such as the one for __attribute__((init_priority(X)))
which must define a class and top-level variable for attribute testing (a
requirement of this particular attribute) rather than the simple
'void f()' used for most tests.


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@30491 8cc4aa7f-3514-0410-904f-f2cc9021211c