Flushed out details of IrradianceCache structure - Looked up "A Ray Tracing Solution for Diffuse Interreflection" by Ward et al. which helped a lot. - Separated out octree node structure to it's own file (it just got too complex). - Filled in details for equations described in Jensen book and Ward's paper. - IrradianceCache is now complete but UNTESTED and un-integrated into photon map.
The irradiance cache determines when a new irradiance sample needs to be calculated from scratch and when it can instead be computed as a weighted sum of previous cached values.  This saves an immense amount of time during the Final Gather phase and, once tested and working, should make debugging the Final Gather easier.

git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/soc2009/advancedlighting@32849 8cc4aa7f-3514-0410-904f-f2cc9021211c